+++ src/main/java/com/sprint/mission/discodeit/dto/AuthLogin/AuthLoginReponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.AuthLogin;
+package com.sprint.mission.discodeit.dto.AuthLogin;
 
 import java.util.UUID;
