+++ src/main/java/com/sprint/mission/discodeit/dto/Channel/ChannelResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.Channel;
+package com.sprint.mission.discodeit.dto.Channel;
 
 import com.sprint.mission.discodeit.entity.ChannelType;
 import java.time.Instant;
