+++ src/main/java/com/sprint/mission/discodeit/dto/Channel/PublicChannelCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.Channel;
+package com.sprint.mission.discodeit.dto.Channel;
 
 import lombok.Builder;
