+++ src/main/java/com/sprint/mission/discodeit/dto/ReadStatus/ReadStatusCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.ReadStatus;
+package com.sprint.mission.discodeit.dto.ReadStatus;
 
 import java.time.Instant;
 import java.util.UUID;
