+++ src/main/java/com/sprint/mission/discodeit/dto/UserStatus/UserStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.UserStatus;
+package com.sprint.mission.discodeit.dto.UserStatus;
 
 import java.time.Instant;
 import java.util.UUID;
