+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.service.dto.AuthLogin.AuthLoginReponse;
-import com.sprint.mission.discodeit.service.dto.AuthLogin.AuthLoginRequest;
+import com.sprint.mission.discodeit.dto.AuthLogin.AuthLoginReponse;
+import com.sprint.mission.discodeit.dto.AuthLogin.AuthLoginRequest;
 
 public interface AuthService {
     public AuthLoginReponse login(AuthLoginRequest request);
