+++ src/main/java/com/sprint/mission/discodeit/service/BinaryContentService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
+import com.sprint.mission.discodeit.dto.BinaryContent.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
-import com.sprint.mission.discodeit.service.dto.BinaryContent.BinaryContentCreateRequest;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
