+++ src/main/java/com/sprint/mission/discodeit/service/MessageService.java
@@ -1,9 +1,9 @@
 package com.sprint.mission.discodeit.service;
 
+import com.sprint.mission.discodeit.dto.BinaryContent.BinaryContentCreateRequest;
+import com.sprint.mission.discodeit.dto.Message.MessageCreateRequest;
+import com.sprint.mission.discodeit.dto.Message.MessageUpdateRequest;
 import com.sprint.mission.discodeit.entity.Message;
-import com.sprint.mission.discodeit.service.dto.BinaryContent.BinaryContentCreateRequest;
-import com.sprint.mission.discodeit.service.dto.Message.MessageCreateRequest;
-import com.sprint.mission.discodeit.service.dto.Message.MessageUpdateRequest;
 import java.util.List;
 import java.util.UUID;
