+++ src/main/java/com/sprint/mission/discodeit/service/ReadStatusService.java
@@ -1,8 +1,8 @@
 package com.sprint.mission.discodeit.service;
 
+import com.sprint.mission.discodeit.dto.ReadStatus.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.ReadStatus.ReadStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.ReadStatus;
-import com.sprint.mission.discodeit.service.dto.ReadStatus.ReadStatusCreateRequest;
-import com.sprint.mission.discodeit.service.dto.ReadStatus.ReadStatusUpdateRequest;
 import java.util.List;
 import java.util.UUID;
