+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,10 +1,10 @@
 package com.sprint.mission.discodeit.service;
 
+import com.sprint.mission.discodeit.dto.User.UserCreateRequest;
+import com.sprint.mission.discodeit.dto.User.UserFindRequest;
+import com.sprint.mission.discodeit.dto.User.UserResponse;
+import com.sprint.mission.discodeit.dto.User.UserUpdateRequest;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.service.dto.User.UserCreateRequest;
-import com.sprint.mission.discodeit.service.dto.User.UserFindRequest;
-import com.sprint.mission.discodeit.service.dto.User.UserResponse;
-import com.sprint.mission.discodeit.service.dto.User.UserUpdateRequest;
 import java.util.List;
 import java.util.UUID;
