+++ src/main/java/com/sprint/mission/discodeit/service/basic/BaiscUserStatusService.java
@@ -1,11 +1,11 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.UserStatus.UserStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.UserStatus.UserStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserStatusService;
-import com.sprint.mission.discodeit.service.dto.UserStatus.UserStatusCreateRequest;
-import com.sprint.mission.discodeit.service.dto.UserStatus.UserStatusUpdateRequest;
 import java.time.Instant;
 import java.util.List;
 import java.util.NoSuchElementException;
